# api/docs/swagger.yaml
basePath: /api/v1/auth
definitions:
  api.AdminCreateUserRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
      password:
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - customer
        - employee
        - admin
    required:
    - email
    - password
    - role
    type: object
  api.AdminUpdateUserRequest:
    properties:
      active:
        type: boolean
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
    type: object
  api.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  api.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  api.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
    type: object
  api.UpdateUserRequest:
    properties:
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
    type: object
  api.UserResponse:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  middleware.HttpError:
    properties:
      code:
        type: integer
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  models.UserRole:
    enum:
    - customer
    - employee
    - admin
    type: string
    x-enum-varnames:
    - RoleCustomer
    - RoleEmployee
    - RoleAdmin
host: localhost:8080
info:
  contact: {}
  description: Authentication and authorization service for the PRPO project
  title: Auth API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      operationId: Login
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      summary: Login user
      tags:
      - auth
  /me:
    get:
      consumes:
      - application/json
      description: Get information about the currently authenticated user
      operationId: GetCurrentUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update information about the currently authenticated user
      operationId: UpdateCurrentUser
      parameters:
      - description: User update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - auth
  /me/password:
    put:
      consumes:
      - application/json
      description: Change password for the currently authenticated user
      operationId: ChangePassword
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Use refresh token to get a new access token
      operationId: RefreshToken
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      summary: Refresh access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      operationId: Register
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      summary: Register a new user
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: List all users (admin endpoint)
      operationId: UsersList
      parameters:
      - default: 10
        description: Limit the number of responses
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset the first response
        in: query
        name: offset
        type: integer
      - description: Sort results
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with any role (admin only)
      operationId: AdminCreateUser
      parameters:
      - description: User creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Create user (admin)
      tags:
      - users
  /users/{userID}:
    delete:
      consumes:
      - application/json
      description: Delete a specific user (admin endpoint)
      operationId: UsersDelete
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a specific user by ID (admin endpoint)
      operationId: UsersShow
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a specific user (admin endpoint)
      operationId: UsersUpdate
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: User update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /verify:
    post:
      consumes:
      - application/json
      description: Verify a JWT token and return user information
      operationId: VerifyToken
      parameters:
      - description: Token to verify
        in: body
        name: token
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.HttpError'
      summary: Verify JWT token
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"